handle event stream with rest
