+++ pkg/client-sdk/ark_sdk.go
@@ -6,7 +6,6 @@ import (
 	"encoding/hex"
 	"errors"
 	"fmt"
-	"io"
 	"math"
 	"strings"
 	"sync"
@@ -15,8 +14,6 @@ import (
 	arkv1 "github.com/ark-network/ark/api-spec/protobuf/gen/ark/v1"
 	"github.com/ark-network/ark/common"
 	"github.com/ark-network/ark/common/tree"
-	"github.com/btcsuite/btcd/btcec/v2/schnorr"
-	"github.com/btcsuite/btcd/chaincfg/chainhash"
 	"github.com/decred/dcrd/dcrec/secp256k1/v4"
 	log "github.com/sirupsen/logrus"
 	"github.com/vulpemventures/go-elements/address"
@@ -988,255 +985,6 @@ func (a *arkClient) coinSelectOnchain(
 	}
 
 	return utxos, delayedUtxos, selectedAmount - targetAmount, nil
-}
-
-func (a *arkClient) handleRoundStream(
-	ctx context.Context,
-	paymentID string,
-	vtxosToSign []vtxo,
-	receivers []*arkv1.Output,
-) (poolTxID string, err error) {
-	stream, err := a.innerClient.getEventStream(ctx, &arkv1.GetEventStreamRequest{})
-	if err != nil {
-		return "", err
-	}
-
-	var pingStop func()
-	pingReq := &arkv1.PingRequest{
-		PaymentId: paymentID,
-	}
-	for pingStop == nil {
-		pingStop = a.ping(ctx, pingReq)
-	}
-
-	defer pingStop()
-
-	for {
-		event, err := stream.Recv()
-		if err == io.EOF {
-			break
-		}
-		if err != nil {
-			return "", err
-		}
-
-		if e := event.GetRoundFailed(); e != nil {
-			pingStop()
-			return "", fmt.Errorf("round failed: %s", e.GetReason())
-		}
-
-		if e := event.GetRoundFinalization(); e != nil {
-			// stop pinging as soon as we receive some forfeit txs
-			pingStop()
-			log.Infoln("round finalization started")
-
-			poolTx := e.GetPoolTx()
-			ptx, err := psetv2.NewPsetFromBase64(poolTx)
-			if err != nil {
-				return "", err
-			}
-
-			congestionTree, err := toCongestionTree(e.GetCongestionTree())
-			if err != nil {
-				return "", err
-			}
-
-			connectors := e.GetConnectors()
-
-			aspPubkey, err := secp256k1.ParsePubKey(a.aspPubKey)
-			if err != nil {
-				return "", err
-			}
-
-			if !isOnchainOnly(receivers) {
-				// validate the congestion tree
-				if err := tree.ValidateCongestionTree(
-					congestionTree, poolTx, aspPubkey, int64(a.roundLifeTime),
-				); err != nil {
-					return "", err
-				}
-			}
-
-			if err := common.ValidateConnectors(poolTx, connectors); err != nil {
-				return "", err
-			}
-
-			for _, receiver := range receivers {
-				isOnChain, onchainScript, userPubkey, err := decodeReceiverAddress(
-					receiver.Address,
-				)
-				if err != nil {
-					return "", err
-				}
-
-				if isOnChain {
-					// collaborative exit case
-					// search for the output in the pool tx
-					found := false
-					for _, output := range ptx.Outputs {
-						if bytes.Equal(output.Script, onchainScript) {
-							if output.Value != receiver.Amount {
-								return "", fmt.Errorf(
-									"invalid collaborative exit output amount: got %d, want %d",
-									output.Value, receiver.Amount,
-								)
-							}
-
-							found = true
-							break
-						}
-					}
-
-					if !found {
-						return "", fmt.Errorf(
-							"collaborative exit output not found: %s", receiver.Address,
-						)
-					}
-
-					continue
-				}
-
-				// off-chain send case
-				// search for the output in congestion tree
-				found := false
-
-				// compute the receiver output taproot key
-				outputTapKey, _, err := computeVtxoTaprootScript(
-					userPubkey, aspPubkey, uint(a.unilateralExitDelay),
-				)
-				if err != nil {
-					return "", err
-				}
-
-				leaves := congestionTree.Leaves()
-				for _, leaf := range leaves {
-					tx, err := psetv2.NewPsetFromBase64(leaf.Tx)
-					if err != nil {
-						return "", err
-					}
-
-					for _, output := range tx.Outputs {
-						if len(output.Script) == 0 {
-							continue
-						}
-						if bytes.Equal(
-							output.Script[2:], schnorr.SerializePubKey(outputTapKey),
-						) {
-							if output.Value != receiver.Amount {
-								continue
-							}
-
-							found = true
-							break
-						}
-					}
-
-					if found {
-						break
-					}
-				}
-
-				if !found {
-					return "", fmt.Errorf(
-						"off-chain send output not found: %s", receiver.Address,
-					)
-				}
-			}
-
-			log.Infoln("congestion tree validated")
-
-			forfeits := e.GetForfeitTxs()
-			signedForfeits := make([]string, 0)
-
-			log.Infoln("signing forfeit txs... ")
-
-			connectorsTxids := make([]string, 0, len(connectors))
-			for _, connector := range connectors {
-				p, _ := psetv2.NewPsetFromBase64(connector)
-				utx, _ := p.UnsignedTx()
-				txid := utx.TxHash().String()
-
-				connectorsTxids = append(connectorsTxids, txid)
-			}
-
-			for _, forfeit := range forfeits {
-				pset, err := psetv2.NewPsetFromBase64(forfeit)
-				if err != nil {
-					return "", err
-				}
-
-				for _, input := range pset.Inputs {
-					inputTxid := chainhash.Hash(input.PreviousTxid).String()
-
-					for _, coin := range vtxosToSign {
-						// check if it contains one of the input to sign
-						if inputTxid == coin.txid {
-							// verify that the connector is in the connectors list
-							connectorTxid := chainhash.Hash(pset.Inputs[0].PreviousTxid).String()
-							connectorFound := false
-							for _, txid := range connectorsTxids {
-								if txid == connectorTxid {
-									connectorFound = true
-									break
-								}
-							}
-
-							if !connectorFound {
-								return "", fmt.Errorf("connector txid %s not found in the connectors list", connectorTxid)
-							}
-
-							_, onchainAddr, _, err := getAddress(
-								a.wallet.PubKeySerializeCompressed(), a.aspPubKey, int64(a.unilateralExitDelay), a.net,
-							)
-							if err != nil {
-								return "", err
-							}
-
-							if err := a.wallet.SignPsetForAddress(pset, onchainAddr); err != nil {
-								return "", err
-							}
-
-							signedPset, err := pset.ToBase64()
-							if err != nil {
-								return "", err
-							}
-
-							signedForfeits = append(signedForfeits, signedPset)
-						}
-					}
-				}
-			}
-
-			// if no forfeit txs have been signed, start pinging again and wait for the next round
-			if len(signedForfeits) == 0 {
-				log.Infoln("no forfeit txs to sign, waiting for the next round...")
-				pingStop = nil
-				for pingStop == nil {
-					pingStop = a.ping(ctx, pingReq)
-				}
-				continue
-			}
-
-			log.Infof("%d signed", len(signedForfeits))
-			log.Infoln("finalizing payment... ")
-			_, err = a.innerClient.finalizePayment(ctx, &arkv1.FinalizePaymentRequest{
-				SignedForfeitTxs: signedForfeits,
-			})
-			if err != nil {
-				return "", err
-			}
-			log.Infoln("done.")
-			log.Infoln("waiting for round finalization...")
-
-			continue
-		}
-
-		if event.GetRoundFinalized() != nil {
-			return event.GetRoundFinalized().GetPoolTxid(), nil
-		}
-	}
-
-	return "", fmt.Errorf("stream closed unexpectedly")
 }
 
 func (a *arkClient) ForceRedeem(ctx context.Context) error {
